+++ src/components/panels/zones/ZonesPanel.tsx
@@ -2,6 +2,7 @@ import "../advanced/AdvancedPanel.css";
 import type { Settings } from "../../../store";
 import FailsafeSection from "./FailsafeSection";
 import CustomStopZoneSection from "./CustomStopZoneSection";
+import ProcessListSection from "./ProcessListSection";
 
 // TODO: This still entirely needs to get done. I've just moved everything over here. If you are a back end dev, please be careful changing UI things xD
 // TODO: Custom Stop zones should be like Sequence clicking where you can add as many as you want in a list.
@@ -14,22 +15,29 @@ interface Props {
 
 export default function ZonesPanel({ settings, update, showInfo }: Props) {
   return (
-    <div className="adv-panel adv-panel-text">
-      <div className="adv-row">
-        <FailsafeSection
-          settings={settings}
-          update={update}
-          showInfo={showInfo}
-        />
-      </div>
-      <div className="adv-columns">
-        <div className="adv-col">
+    <div className="adv-panel adv-panel-text adv-panel--zones">
+      <div className="adv-zones-row">
+        <div className="adv-zones-left">
+          <FailsafeSection
+            settings={settings}
+            update={update}
+            showInfo={showInfo}
+          />
+        </div>
+        <div className="adv-zones-right">
           <CustomStopZoneSection
             settings={settings}
             update={update}
             showInfo={showInfo}
           />
         </div>
+      </div>
+      <div className="adv-row adv-row--grow">
+        <ProcessListSection
+          settings={settings}
+          update={update}
+          showInfo={showInfo}
+        />
       </div>
     </div>
   );